--- app/src/main/java/vn/edu/usth/onlinemusicplayer/model/SongsList.java
@@ -1,36 +1,0 @@
-package vn.edu.usth.onlinemusicplayer.model;
-
-public class SongsList {
-
-    private String title;
-    private String subTitle;
-    private String path;
-
-    public String getPath() {
-        return path;
-    }
-
-    public void setPath(String path) {
-        this.path = path;
-    }
-
-    public SongsList(String title, String subTitle, String path) {
-        this.title = title;
-        this.subTitle = subTitle;
-        this.path = path;
-
-    }
-
-    public String getTitle() {
-        return title;
-    }
-
-    public void setTitle(String title) {
-        this.title = title;
-    }
-
-    public String getSubTitle() {
-        return subTitle;
-    }
-
-}
