+++ app/src/main/res/drawable/ic_play_circle.xml
