--- app/src/main/res/layout/activity_audio_player.xml
@@ -1,77 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.drawerlayout.widget.DrawerLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/drawer_layout"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".activity.AudioPlayerActivity"
-    tools:ignore="MissingClass">
-
-    <RelativeLayout
-        android:id="@+id/content_frame"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent">
-
-        <androidx.appcompat.widget.Toolbar
-            android:id="@+id/toolbar"
-            android:layout_width="match_parent"
-            android:layout_height="?attr/actionBarSize"
-            android:background="?attr/colorPrimary"
-            android:theme="@style/ThemeOverlay.AppCompat.ActionBar" />
-
-        <LinearLayout
-            android:id="@+id/ll_include_controls"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_below="@id/toolbar">
-
-            <include layout="@layout/player_layout" />
-
-        </LinearLayout>
-
-        <LinearLayout
-            android:id="@+id/ll_tab_layout"
-            android:layout_width="match_parent"
-            android:layout_height="?attr/actionBarSize"
-            android:layout_below="@id/ll_include_controls"
-            android:background="@color/colorPrimary">
-
-            <com.google.android.material.tabs.TabLayout
-                android:id="@+id/tabs"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:elevation="10dp"
-                app:tabBackground="@color/colorPrimary"
-                app:tabGravity="fill"
-                app:tabIndicatorColor="@color/colorPrimary"
-                app:tabMode="fixed"
-                app:tabSelectedTextColor="@color/text_color"
-                app:tabTextColor="@color/off_color" />
-        </LinearLayout>
-
-
-        <androidx.viewpager.widget.ViewPager
-            android:id="@+id/songs_viewpager"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_below="@id/ll_tab_layout"
-            app:layout_behavior="@string/appbar_scrolling_view_behavior" />
-
-        <com.google.android.material.floatingactionbutton.FloatingActionButton
-            android:id="@+id/btn_refresh"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_alignParentBottom="true"
-            android:layout_alignParentEnd="true"
-            android:layout_marginBottom="15dp"
-            android:layout_marginEnd="15dp"
-            android:focusable="true"
-            android:src="@drawable/ic_refresh"
-            app:backgroundTint="@color/colorPrimary"
-            app:layout_anchor="@id/songs_viewpager"
-            android:layout_alignParentRight="true"
-            android:layout_marginRight="15dp" />
-
-    </RelativeLayout>
-</androidx.drawerlayout.widget.DrawerLayout>
