+++ app/src/main/res/layout/item_song.xml
@@ -90,6 +90,6 @@
         app:layout_constraintBottom_toBottomOf="@+id/cl_song_item"
         app:layout_constraintRight_toRightOf="parent"
         app:layout_constraintTop_toTopOf="@+id/cl_song_item"
-        app:srcCompat="@drawable/ic_menu" />
+        app:srcCompat="@drawable/ic_play_circle" />
 </androidx.constraintlayout.widget.ConstraintLayout>
 </androidx.cardview.widget.CardView>
