--- app/src/main/res/layout/player_layout.xml
@@ -1,99 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:layout_width="match_parent"
-    android:layout_height="100dp"
-    android:background="@color/colorPrimary">
-
-    <LinearLayout
-        android:id="@+id/ll_player_controls"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="10dp"
-        android:gravity="center_horizontal"
-        android:orientation="horizontal">
-
-        <ImageButton
-            android:id="@+id/img_btn_replay"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="10dp"
-            android:backgroundTint="@color/colorPrimary"
-            android:src="@drawable/ic_undo"
-            android:layout_marginLeft="10dp" />
-
-        <ImageButton
-            android:id="@+id/img_btn_previous"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="30dp"
-            android:backgroundTint="@color/colorPrimary"
-            android:src="@drawable/ic_arrow_back"
-            android:layout_marginLeft="30dp" />
-
-        <ImageButton
-            android:id="@+id/img_btn_play"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="10dp"
-            android:backgroundTint="@color/colorPrimary"
-            android:src="@drawable/ic_baseline_play_circle_outline_24"
-            android:layout_marginLeft="10dp" />
-
-        <ImageButton
-            android:id="@+id/img_btn_next"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="10dp"
-            android:backgroundTint="@color/colorPrimary"
-            android:src="@drawable/ic_baseline_arrow_forward_ios_24"
-            android:layout_marginLeft="10dp" />
-
-        <ImageButton
-            android:id="@+id/img_btn_setting"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="30dp"
-            android:backgroundTint="@color/colorPrimary"
-            android:src="@drawable/ic_loop"
-            android:layout_marginLeft="30dp" />
-    </LinearLayout>
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_below="@id/ll_player_controls"
-        android:layout_marginTop="10dp"
-        android:gravity="center_horizontal"
-        android:orientation="horizontal">
-
-        <TextView
-            android:id="@+id/tv_current_time"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="@string/start_time"
-            android:textColor="@color/text_color"
-            android:textSize="15sp"
-            android:textStyle="bold" />
-
-        <SeekBar
-            android:id="@+id/seekbar_controller"
-            android:layout_width="250dp"
-            android:layout_height="wrap_content"
-            android:layout_gravity="center"
-            android:backgroundTint="@color/text_color"
-            android:fitsSystemWindows="true"
-            android:progressTint="@color/text_color"
-            android:thumbTint="@color/text_color" />
-
-        <TextView
-            android:id="@+id/tv_total_time"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="@string/final_time"
-            android:textColor="@color/text_color"
-            android:textSize="15sp"
-            android:textStyle="bold" />
-    </LinearLayout>
-
-</RelativeLayout>
